+++ go.mod
@@ -1,13 +1,7 @@
 module github.com/caigwatkin/slate
 
 require (
-	cloud.google.com/go v0.28.0 // indirect
+	github.com/caigwatkin/go v0.0.0-20180924193504-fdae37a32f86
 	github.com/go-chi/chi v3.3.3+incompatible
-	github.com/google/uuid v1.0.0
 	github.com/pkg/errors v0.8.0
-	golang.org/x/net v0.0.0-20180921000356-2f5d2388922f // indirect
-	golang.org/x/oauth2 v0.0.0-20180821212333-d2e6202438be
-	golang.org/x/sync v0.0.0-20180314180146-1d60e4601c6f // indirect
-	google.golang.org/api v0.0.0-20180921000521-920bb1beccf7
-	google.golang.org/appengine v1.2.0 // indirect
 )
