+++ scripts/go/start.sh
